Fix datatypes for linux execution

+++ simpleRT/model3d.pyx
@@ -14,7 +14,7 @@ from simpleRT.helpers.objloader import OBJ
 
 cdef class Face3D: 
     def __init__(self, tuple objFace = None):
-        self._vertices = np.empty((3,),dtype='int')
+        self._vertices = np.empty((3,),dtype=np.int32)
         self.vertices = self._vertices
         if objFace is None:
             self.normal_idx = -1
